refactor: eliminate code duplication and improve code quality (#15)
* refactor: eliminate code duplication and improve code quality

- Extract shared harness runner (runner.rs) eliminating ~85% duplication
  between HostHarness and KataHarness
- Split daemon.rs (238→122+128) and client.rs (208→159+59) under 200-line limit
- Merge identical is_fatal_token_error/is_fatal_request_error into
  is_fatal_api_error
- Delete dead ContainerBoot error variant
- Consolidate parse tests into parse_tests.rs, remove duplicated tests
- Migration 006: widen tokens_used and duration_ms from INTEGER to BIGINT
- Update Rust types i32→i64 for tokens_used/duration_ms across main-app
- Regenerate sqlx-data.json

* fix: restore secret forwarding into Kata container

The refactored runner.rs applied secrets via cmd.env() generically,
which sets environment on the Docker CLI process but doesn't propagate
them inside the container. Move secret application into each harness's
build_cmd closure: host uses cmd.env(), kata uses -e KEY=VALUE flags.

Remove secrets field from RunnerEnv since it's no longer read there.

+++ main-app/src/services/work_runs/repository/work_runs.rs
@@ -155,7 +155,7 @@ impl WorkRunsRepository {
 pub struct SetResultParams<'a> {
     pub pr_url: &'a str,
     pub exit_code: i32,
-    pub tokens_used: i32,
-    pub duration_ms: i32,
+    pub tokens_used: i64,
+    pub duration_ms: i64,
     pub status: WorkRunStatus,
 }
